conversation-history/pyproject.toml
[tool.poetry]
name = "conversation-history"
version = "0.1.0"
description = ""
authors = ["Ryuichi Inagaki <ryuichi.inagaki.24@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
lark-parser = "^0.8.5"
google-cloud-logging = "^1.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
main = "main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
